// frontend/src/lib/api/models/SectionDetail.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SectionDetail = {
    id: string;
    courseModuleId: string;
    title: string;
    type: string;
    publishStatus: string;
    lecturerNotes: (string | null);
    updatedAt: string;
};